refactor: separate datasets in files
