+++ src/datasets/categories.ts
@@ -1,0 +1,49 @@
+export const categories = [
+    {
+      name: "Clothing",
+      products: [
+        {
+          name: "Orange pants",
+        },
+        {
+          name: "Black pants",
+        },
+        {
+          name: "Red pants",
+        },
+        {
+          name: "Blue T-Shirt",
+        },
+      ],
+    },
+    {
+      name: "Female",
+      products: [
+        {
+          name: "Orange pants",
+        },
+        {
+          name: "Red pants",
+        },
+      ],
+    },
+    {
+      name: "Male",
+      products: [
+        {
+          name: "Black pants",
+        },
+        {
+          name: "Blue T-Shirt",
+        },
+      ],
+    },
+    {
+      name: "Kids",
+      products: [
+        {
+          name: "Blue T-Shirt",
+        },
+      ],
+    },
+  ];
