+++ src/datasets/pictures.ts
@@ -1,0 +1,30 @@
+export const pictures = [
+    {
+      source:
+        "https://unsplash.com/pt-br/fotografias/SdR2wW-v4PE?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      product: {
+        name: "Orange pants",
+      },
+    },
+    {
+      source:
+        "https://unsplash.com/pt-br/fotografias/ylHcWlrMlzs?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      product: {
+        name: "Black pants",
+      },
+    },
+    {
+      source:
+        "https://unsplash.com/pt-br/fotografias/3we24FcjVAk?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      product: {
+        name: "Red pants",
+      },
+    },
+    {
+      source:
+        "https://unsplash.com/pt-br/fotografias/BN760VSO8yM?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      product: {
+        name: "Blue T-Shirt",
+      },
+    },
+  ];
