+++ src/datasets/products.ts
@@ -1,0 +1,62 @@
+export const products = [
+    {
+      id: "P001",
+      name: "Orange pants",
+      description: "Modern Orange pants",
+      price: "130.0",
+      categories: [{ name: "Clothing" }, { name: "Female" }],
+      picture: {
+        source:
+          "https://unsplash.com/pt-br/fotografias/SdR2wW-v4PE?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      },
+      color: "ORANGE",
+    },
+    {
+      id: "P002",
+      name: "Orange Wool Sweater",
+      description: "Confy Orange Wool Sweater",
+      price: "100.0",
+      categories: [{ name: "Clothing" }, { name: "Female" }],
+      picture: {
+        source:
+          "https://unsplash.com/pt-br/fotografias/SdR2wW-v4PE?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      },
+      color: "ORANGE",
+    },
+    {
+      id: "P003",
+      name: "Black pants",
+      description: "Social Black pants",
+      price: "125.0",
+      categories: [{ name: "Clothing" }, { name: "Male" }],
+      picture: {
+        source:
+          "https://unsplash.com/pt-br/fotografias/ylHcWlrMlzs?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      },
+      color: "BLACK",
+    },
+    {
+      id: "P004",
+      name: "Red pants",
+      description: "Confy Red pants",
+      price: "150.0",
+      categories: [{ name: "Clothing" }, { name: "Female" }],
+      picture: {
+        source:
+          "https://unsplash.com/pt-br/fotografias/3we24FcjVAk?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      },
+      color: "RED",
+    },
+    {
+      id: "P005",
+      name: "Blue T-Shirt",
+      description: "Confy Blue T-Shirt",
+      price: "40.0",
+      categories: [{ name: "Clothing" }, { name: "Male" }, { name: "Kids" }],
+      picture: {
+        source:
+          "https://unsplash.com/pt-br/fotografias/BN760VSO8yM?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText",
+      },
+      color: "BLUE",
+    },
+  ];
